split up DashboardView
